+++ mepo.d/utest/test_mepo_commands.py
@@ -10,10 +10,12 @@ from io import StringIO
 
 from input import args
 
-from command.init   import init   as mepo_init
-from command.clone  import clone  as mepo_clone
-from command.list   import list   as mepo_list
-from command.status import status as mepo_status
+from command.init    import init    as mepo_init
+from command.clone   import clone   as mepo_clone
+from command.list    import list    as mepo_list
+from command.status  import status  as mepo_status
+from command.compare import compare as mepo_compare
+from command.develop import develop as mepo_develop
 
 class TestMepoCommands(unittest.TestCase):
 
@@ -34,13 +36,13 @@ class TestMepoCommands(unittest.TestCase):
         cls.input_dir = os.path.join(THIS_DIR, 'input')
         cls.output_dir = os.path.join(THIS_DIR, 'output')
         cls.fixture = 'GEOSfvdycore'
-        cls.tag = 'v1.2.7'
+        cls.tag = 'v1.5.0'
         cls.tmpdir = os.path.join(THIS_DIR, 'tmp')
         cls.fixture_dir = os.path.join(cls.tmpdir, cls.fixture)
         if os.path.isdir(cls.fixture_dir):
             shutil.rmtree(cls.fixture_dir)
         cls.__checkout_fixture()
-        cls.__copy_config_file()
+        #cls.__copy_config_file()
         args.config = 'components.yaml'
         args.style = 'prefix'
         os.chdir(cls.fixture_dir)
@@ -50,6 +52,10 @@ class TestMepoCommands(unittest.TestCase):
         args.branch = None
         args.directory = None
         mepo_clone.run(args)
+        # In order to better test compare, we need to do *something*
+        args.comp_name = ['env','cmake','fvdycore']
+        args.quiet = False
+        mepo_develop.run(args)
 
     def setUp(self):
         pass
@@ -65,9 +71,30 @@ class TestMepoCommands(unittest.TestCase):
     def test_status(self):
         sys.stdout = output = StringIO()
         args.ignore_permissions=False
+        args.nocolor=True
         mepo_status.run(args)
         sys.stdout = sys.__stdout__
         with open(os.path.join(self.__class__.output_dir, 'status_output.txt'), 'r') as fin:
+            saved_output = fin.read()
+        self.assertEqual(output.getvalue(), saved_output)
+
+    def test_compare_brief(self):
+        sys.stdout = output = StringIO()
+        args.all=False
+        args.nocolor=True
+        mepo_compare.run(args)
+        sys.stdout = sys.__stdout__
+        with open(os.path.join(self.__class__.output_dir, 'compare_brief_output.txt'), 'r') as fin:
+            saved_output = fin.read()
+        self.assertEqual(output.getvalue(), saved_output)
+
+    def test_compare_full(self):
+        sys.stdout = output = StringIO()
+        args.all=True
+        args.nocolor=True
+        mepo_compare.run(args)
+        sys.stdout = sys.__stdout__
+        with open(os.path.join(self.__class__.output_dir, 'compare_full_output.txt'), 'r') as fin:
             saved_output = fin.read()
         self.assertEqual(output.getvalue(), saved_output)