+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Changed the default behavior of `compare` to only show differing repos. Use `--all` to see all repos
+
 ### Removed
 
 ## [1.40.0] - 2022-01-12